test(tui): split app/tests.rs by App behavior contract
Break the 2055-line `rinkaku-tui/src/app/tests.rs` (already extracted
via `#[path = "tests.rs"] mod tests;`) into `app/tests/*.rs` grouped
by the App behavior each test pins:

- `basics` — startup defaults, `Quit`, `ToggleOrder`, cursor `Down`,
  status clearing, `selected_detail`
- `source_screen` — `Source`/`Back` transitions and all
  `Screen::Source` scroll behaviors (ADR 0026)
- `right_pane` — right-pane toggles (`ToggleDiff`,
  `ToggleBlastRadius`), `selected_diff_target`/`selected_diff_focus`
- `focus` — `Focus::Tree`/`Right` defaults and transitions,
  `FocusLeft`, `PendingGoto` scroll preservation
- `scroll_reset` — blanket `right_pane_scroll` reset invariants
- `help_overlay` — `ToggleHelp` open/close, help scroll
- `jump` — `selected_symbol_id`, jump popup, `PendingPrefix::G`
  bookkeeping, `JumpBack`/`JumpForward` jumplist (ADR 0022)

Shared fixtures (`empty_report`, `report_with_one_symbol`,
`report_with_two_directories`, etc.) move into `tests/mod.rs`.
`app/mod.rs`'s `#[path]` changes from `"tests.rs"` to
`"tests/mod.rs"`; the production body is untouched. Test count
unchanged (524 rinkaku_tui tests still pass).
